move CompactReceipt into compact.rs
