+++ risc0/zkvm/src/host/receipt.rs
@@ -14,6 +14,7 @@
 
 //! Manages the output and cryptographic data for a proven computation.
 
+pub(crate) mod compact;
 pub(crate) mod composite;
 pub(crate) mod segment;
 
@@ -21,11 +22,7 @@ use alloc::{collections::BTreeMap, string::String, vec, vec::Vec};
 use core::fmt::Debug;
 
 use anyhow::Result;
-use risc0_circuit_recursion::control_id::{ALLOWED_CONTROL_ROOT, BN254_CONTROL_ID};
 use risc0_core::field::baby_bear::BabyBear;
-use risc0_groth16::{
-    fr_from_hex_string, split_digest, verifier::prepared_verifying_key, Seal, Verifier,
-};
 use risc0_zkp::{
     core::{
         digest::Digest,
@@ -44,7 +41,7 @@ use crate::{
     Assumptions, MaybePruned, Output, ReceiptClaim,
 };
 
-pub use self::{composite::CompositeReceipt, segment::SegmentReceipt};
+pub use self::{compact::CompactReceipt, composite::CompositeReceipt, segment::SegmentReceipt};
 pub use super::recursion::SuccinctReceipt;
 
 /// A receipt attesting to the execution of a guest program.
@@ -362,46 +359,6 @@ impl InnerReceipt {
             InnerReceipt::Succinct(ref succinct_receipt) => Ok(succinct_receipt.claim.clone()),
             InnerReceipt::Fake { claim } => Ok(claim.clone()),
         }
-    }
-}
-
-/// A receipt composed of a Groth16 over the BN_254 curve
-#[derive(Clone, Debug, Deserialize, Serialize)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct CompactReceipt {
-    /// A Groth16 proof of a zkVM execution with the associated claim.
-    pub seal: Vec<u8>,
-
-    /// [ReceiptClaim] containing information about the execution that this
-    /// receipt proves.
-    pub claim: ReceiptClaim,
-}
-
-impl CompactReceipt {
-    /// Verify the integrity of this receipt, ensuring the claim is attested
-    /// to by the seal.
-    pub fn verify_integrity(&self) -> Result<(), VerificationError> {
-        use hex::FromHex;
-        let (a0, a1) = split_digest(
-            Digest::from_hex(ALLOWED_CONTROL_ROOT)
-                .map_err(|_| VerificationError::ReceiptFormatError)?,
-        )
-        .map_err(|_| VerificationError::ReceiptFormatError)?;
-        let (c0, c1) =
-            split_digest(self.claim.digest()).map_err(|_| VerificationError::ReceiptFormatError)?;
-        let id_p254_hash = fr_from_hex_string(BN254_CONTROL_ID)
-            .map_err(|_| VerificationError::ReceiptFormatError)?;
-        Verifier::new(
-            &Seal::from_vec(&self.seal).map_err(|_| VerificationError::ReceiptFormatError)?,
-            vec![a0, a1, c0, c1, id_p254_hash],
-            prepared_verifying_key().map_err(|_| VerificationError::ReceiptFormatError)?,
-        )
-        .map_err(|_| VerificationError::ReceiptFormatError)?
-        .verify()
-        .map_err(|_| VerificationError::InvalidProof)?;
-
-        // Everything passed
-        Ok(())
     }
 }
